Update Spacy Version
Fixes memory leak found in Spacy 2.1.8

"This is a small maintenance update that backports a bug fix for a memory leak that'd occur in long-running parsing processes. It's intended for users who can't or don't yet want to upgrade to spaCy v2.2 (e.g. because it requires retraining all the models). If you're able to upgrade, you shouldn't use this version and instead install the latest v2.2." - https://github.com/explosion/spaCy/releases/tag/v2.1.9